Split iOS delete computer regression code
