+++ Packages/iOS/CmuxMobileShell/Tests/CmuxMobileShellTests/MobileShellCompositePairedMacCoalescingTests.swift
@@ -109,92 +109,6 @@ import Testing
         #expect(try await pairedStore.loadAll(stackUserID: "user-1", teamID: "team-a").map(\.macDeviceID) == ["mac-other"])
     }
 
-    @Test func forgettingLastVisibleMacClearsSavedMacHint() async throws {
-        let defaultsSuiteName = "forget-last-mac-hint-\(UUID().uuidString)"
-        let defaults = try #require(UserDefaults(suiteName: defaultsSuiteName))
-        defaults.set(true, forKey: "cmux.mobile.hasKnownPairedMac")
-        defer { defaults.removePersistentDomain(forName: defaultsSuiteName) }
-        let pairedStore = DelayedTeamPairedMacStore(
-            recordsByTeam: [
-                "team-a": [
-                    try Self.pairedMac(
-                        id: "mac-a",
-                        displayName: "Desk Mac",
-                        host: "100.82.214.112",
-                        lastSeenAt: Date(timeIntervalSince1970: 10),
-                        isActive: true
-                    ),
-                ],
-            ],
-            blockedTeams: []
-        )
-        let store = MobileShellComposite(
-            isSignedIn: true,
-            pairedMacStore: pairedStore,
-            identityProvider: StaticIdentityProvider(userID: "user-1"),
-            teamIDProvider: { "team-a" },
-            pairingHintDefaults: defaults
-        )
-        await store.loadPairedMacs()
-        #expect(store.hasKnownPairedMac)
-
-        await store.forgetMac(macDeviceID: "mac-a")
-
-        #expect(store.pairedMacs.isEmpty)
-        #expect(store.displayPairedMacs.isEmpty)
-        #expect(!store.hasKnownPairedMac)
-    }
-
-    @Test func forgetStoredMacRemovesOnlyExactAliasRow() async throws {
-        let defaultsSuiteName = "forget-exact-alias-hint-\(UUID().uuidString)"
-        let defaults = try #require(UserDefaults(suiteName: defaultsSuiteName))
-        defaults.set(true, forKey: "cmux.mobile.hasKnownPairedMac")
-        defer { defaults.removePersistentDomain(forName: defaultsSuiteName) }
-        let pairedStore = DelayedTeamPairedMacStore(
-            recordsByTeam: [
-                "team-a": [
-                    try Self.pairedMac(
-                        id: "mac-old",
-                        displayName: "Lawrence Mac",
-                        host: "100.82.214.112",
-                        lastSeenAt: Date(timeIntervalSince1970: 10),
-                        isActive: false
-                    ),
-                    try Self.pairedMac(
-                        id: "mac-fresh",
-                        displayName: "Lawrence Mac",
-                        host: "100.82.214.112",
-                        lastSeenAt: Date(timeIntervalSince1970: 20),
-                        isActive: true
-                    ),
-                    try Self.pairedMac(
-                        id: "mac-other",
-                        displayName: "Other Mac",
-                        host: "100.82.214.113",
-                        lastSeenAt: Date(timeIntervalSince1970: 30),
-                        isActive: false
-                    ),
-                ],
-            ],
-            blockedTeams: []
-        )
-        let store = MobileShellComposite(
-            isSignedIn: true,
-            pairedMacStore: pairedStore,
-            identityProvider: StaticIdentityProvider(userID: "user-1"),
-            teamIDProvider: { "team-a" },
-            pairingHintDefaults: defaults
-        )
-        await store.loadPairedMacs()
-
-        await store.forgetStoredMac(macDeviceID: "mac-old")
-
-        #expect(try await pairedStore.loadAll(stackUserID: "user-1", teamID: "team-a").map(\.macDeviceID) == ["mac-fresh", "mac-other"])
-        #expect(store.pairedMacs.map(\.macDeviceID) == ["mac-fresh", "mac-other"])
-        #expect(store.displayPairedMacs.map(\.macDeviceID) == ["mac-fresh", "mac-other"])
-        #expect(store.hasKnownPairedMac)
-    }
-
     @Test func presenceRoutesForHiddenDuplicateRefreshOnlyTheEmittingRow() async throws {
         let pairedStore = DelayedTeamPairedMacStore(
             recordsByTeam: [
@@ -393,624 +307,6 @@ import Testing
         #expect(store.displayPairedMacs.map(\.macDeviceID) == ["mac-a", "mac-b"])
     }
 
-    @Test func forgettingMacClearsAnonymousWorkspaceSnapshotOwnedByThatMac() async throws {
-        let pairedStore = DelayedTeamPairedMacStore(
-            recordsByTeam: [
-                "team-a": [
-                    try Self.pairedMac(
-                        id: "mac-a",
-                        displayName: "Lawrence Mac",
-                        host: "100.82.214.112",
-                        lastSeenAt: Date(timeIntervalSince1970: 10),
-                        isActive: false
-                    ),
-                ],
-            ],
-            blockedTeams: []
-        )
-        let store = MobileShellComposite(
-            isSignedIn: true,
-            pairedMacStore: pairedStore,
-            identityProvider: StaticIdentityProvider(userID: "user-1"),
-            teamIDProvider: { "team-a" }
-        )
-        await store.loadPairedMacs()
-        store.setWorkspaceStatesForTesting([
-            MobileShellComposite.foregroundAnonymousKey: MacWorkspaceState(
-                macDeviceID: MobileShellComposite.foregroundAnonymousKey,
-                workspaces: [
-                    MobileWorkspacePreview(
-                        id: "stale-workspace",
-                        macDeviceID: "mac-a",
-                        name: "Stale",
-                        terminals: []
-                    ),
-                ],
-                status: .unavailable
-            ),
-        ], foregroundMacDeviceID: nil)
-        #expect(store.workspaces.map(\.rpcWorkspaceID.rawValue) == ["stale-workspace"])
-
-        await store.forgetMac(macDeviceID: "mac-a")
-
-        #expect(store.pairedMacs.isEmpty)
-        #expect(store.displayPairedMacs.isEmpty)
-        #expect(store.workspaces.isEmpty)
-    }
-
-    @Test func forgettingActiveMacPreservesRemainingMacWorkspaceSnapshot() async throws {
-        let pairedStore = DelayedTeamPairedMacStore(
-            recordsByTeam: [
-                "team-a": [
-                    try Self.pairedMac(
-                        id: "mac-a",
-                        displayName: "Desk Mac",
-                        host: "100.82.214.112",
-                        lastSeenAt: Date(timeIntervalSince1970: 10),
-                        isActive: true
-                    ),
-                    try Self.pairedMac(
-                        id: "mac-b",
-                        displayName: "Laptop Mac",
-                        host: "100.82.214.113",
-                        lastSeenAt: Date(timeIntervalSince1970: 20),
-                        isActive: false
-                    ),
-                ],
-            ],
-            blockedTeams: []
-        )
-        let store = MobileShellComposite(
-            isSignedIn: true,
-            connectionState: .connected,
-            pairedMacStore: pairedStore,
-            identityProvider: StaticIdentityProvider(userID: "user-1"),
-            teamIDProvider: { "team-a" }
-        )
-        await store.loadPairedMacs()
-        store.setWorkspaceStatesForTesting([
-            "mac-a": MacWorkspaceState(
-                macDeviceID: "mac-a",
-                workspaces: [
-                    MobileWorkspacePreview(
-                        id: "deleted-workspace",
-                        macDeviceID: "mac-a",
-                        name: "Deleted",
-                        terminals: []
-                    ),
-                ],
-                status: .connected
-            ),
-            "mac-b": MacWorkspaceState(
-                macDeviceID: "mac-b",
-                workspaces: [
-                    MobileWorkspacePreview(
-                        id: "remaining-workspace",
-                        macDeviceID: "mac-b",
-                        name: "Remaining",
-                        terminals: []
-                    ),
-                ],
-                status: .connected
-            ),
-        ], foregroundMacDeviceID: "mac-a")
-        #expect(store.workspaces.map(\.rpcWorkspaceID.rawValue) == ["deleted-workspace", "remaining-workspace"])
-
-        await store.forgetMac(macDeviceID: "mac-a")
-
-        #expect(store.pairedMacs.map(\.macDeviceID) == ["mac-b"])
-        #expect(store.workspaces.map(\.rpcWorkspaceID.rawValue) == ["remaining-workspace"])
-        #expect(store.connectionState == .disconnected)
-        #expect(store.macConnectionStatus == .unavailable)
-        #expect(store.workspaceListConnectionStatus == .connected)
-        #expect(store.workspaceListConnectedRefreshTargetMacDeviceID() == "mac-b")
-    }
-
-    @Test func forgettingMacFiltersOnlyMatchingRowsFromMixedWorkspaceBucket() async throws {
-        let pairedStore = DelayedTeamPairedMacStore(
-            recordsByTeam: [
-                "team-a": [
-                    try Self.pairedMac(
-                        id: "mac-a",
-                        displayName: "Desk Mac",
-                        host: "100.82.214.112",
-                        lastSeenAt: Date(timeIntervalSince1970: 10),
-                        isActive: false
-                    ),
-                    try Self.pairedMac(
-                        id: "mac-b",
-                        displayName: "Laptop Mac",
-                        host: "100.82.214.113",
-                        lastSeenAt: Date(timeIntervalSince1970: 20),
-                        isActive: true
-                    ),
-                ],
-            ],
-            blockedTeams: []
-        )
-        let store = MobileShellComposite(
-            isSignedIn: true,
-            connectionState: .connected,
-            pairedMacStore: pairedStore,
-            identityProvider: StaticIdentityProvider(userID: "user-1"),
-            teamIDProvider: { "team-a" }
-        )
-        await store.loadPairedMacs()
-        store.setWorkspaceStatesForTesting([
-            MobileShellComposite.foregroundAnonymousKey: MacWorkspaceState(
-                macDeviceID: MobileShellComposite.foregroundAnonymousKey,
-                workspaces: [
-                    MobileWorkspacePreview(
-                        id: "deleted-workspace",
-                        macDeviceID: "mac-a",
-                        name: "Deleted",
-                        terminals: []
-                    ),
-                    MobileWorkspacePreview(
-                        id: "remaining-workspace",
-                        macDeviceID: "mac-b",
-                        name: "Remaining",
-                        terminals: []
-                    ),
-                ],
-                status: .connected
-            ),
-        ], foregroundMacDeviceID: nil)
-
-        await store.forgetMac(macDeviceID: "mac-a")
-
-        #expect(store.pairedMacs.map(\.macDeviceID) == ["mac-b"])
-        #expect(store.workspaces.map(\.rpcWorkspaceID.rawValue) == ["remaining-workspace"])
-        #expect(store.workspaceListConnectionStatus == .connected)
-    }
-
-    @Test func failedForgetRestoresMacVisibilityAndForgottenTombstone() async throws {
-        let pairedStore = DelayedTeamPairedMacStore(
-            recordsByTeam: [
-                "team-a": [
-                    try Self.pairedMac(
-                        id: "mac-a",
-                        displayName: "Desk Mac",
-                        host: "100.82.214.112",
-                        lastSeenAt: Date(timeIntervalSince1970: 10),
-                        isActive: false
-                    ),
-                    try Self.pairedMac(
-                        id: "mac-b",
-                        displayName: "Laptop Mac",
-                        host: "100.82.214.113",
-                        lastSeenAt: Date(timeIntervalSince1970: 20),
-                        isActive: true
-                    ),
-                ],
-            ],
-            blockedTeams: []
-        )
-        await pairedStore.failRemove(macDeviceID: "mac-a")
-        let store = MobileShellComposite(
-            isSignedIn: true,
-            connectionState: .connected,
-            pairedMacStore: pairedStore,
-            identityProvider: StaticIdentityProvider(userID: "user-1"),
-            teamIDProvider: { "team-a" },
-            forgottenMacStore: InMemoryPairedMacForgottenStore()
-        )
-        await store.loadPairedMacs()
-        store.setWorkspaceStatesForTesting([
-            "mac-a": MacWorkspaceState(
-                macDeviceID: "mac-a",
-                workspaces: [
-                    MobileWorkspacePreview(
-                        id: "mac-a-workspace",
-                        macDeviceID: "mac-a",
-                        name: "Desk",
-                        terminals: []
-                    ),
-                ],
-                status: .connected
-            ),
-        ], foregroundMacDeviceID: nil)
-
-        await store.forgetMac(macDeviceID: "mac-a")
-        await store.loadPairedMacs()
-
-        #expect(store.pairedMacs.map(\.macDeviceID) == ["mac-a", "mac-b"])
-        #expect(store.displayPairedMacs.map(\.macDeviceID) == ["mac-a", "mac-b"])
-        #expect(store.workspaces.map(\.rpcWorkspaceID.rawValue) == ["mac-a-workspace"])
-    }
-
-    @Test func forgettingMacSuppressesStaleStoreWriteFromSameSession() async throws {
-        let pairedStore = DelayedTeamPairedMacStore(
-            recordsByTeam: [
-                "team-a": [
-                    try Self.pairedMac(
-                        id: "mac-a",
-                        displayName: "Desk Mac",
-                        host: "100.82.214.112",
-                        lastSeenAt: Date(timeIntervalSince1970: 10),
-                        isActive: true
-                    ),
-                    try Self.pairedMac(
-                        id: "mac-b",
-                        displayName: "Laptop Mac",
-                        host: "100.82.214.113",
-                        lastSeenAt: Date(timeIntervalSince1970: 20),
-                        isActive: false
-                    ),
-                ],
-            ],
-            blockedTeams: []
-        )
-        let store = MobileShellComposite(
-            isSignedIn: true,
-            pairedMacStore: pairedStore,
-            identityProvider: StaticIdentityProvider(userID: "user-1"),
-            teamIDProvider: { "team-a" }
-        )
-        await store.loadPairedMacs()
-
-        await store.forgetMac(macDeviceID: "mac-a")
-        try await pairedStore.upsert(
-            macDeviceID: "mac-a",
-            displayName: "Desk Mac",
-            routes: [try CmxAttachRoute(
-                id: "stale",
-                kind: .tailscale,
-                endpoint: .hostPort(host: "100.82.214.112", port: 50922)
-            )],
-            markActive: true,
-            stackUserID: "user-1",
-            teamID: "team-a",
-            now: Date(timeIntervalSince1970: 30)
-        )
-
-        await store.loadPairedMacs()
-
-        #expect(store.pairedMacs.map(\.macDeviceID) == ["mac-b"])
-        #expect(store.displayPairedMacs.map(\.macDeviceID) == ["mac-b"])
-        #expect(await store.secondaryAggregationCandidateMacIDsForTesting() == ["mac-b"])
-    }
-
-    @Test func forgettingMacSuppressesStaleStoreWriteAfterRelaunch() async throws {
-        let suiteName = "forgotten-mac-relaunch-\(UUID().uuidString)"
-        let forgottenStore = UserDefaultsPairedMacForgottenStore(suiteName: suiteName)
-        defer { UserDefaults(suiteName: suiteName)?.removePersistentDomain(forName: suiteName) }
-        let pairedStore = DelayedTeamPairedMacStore(
-            recordsByTeam: [
-                "team-a": [
-                    try Self.pairedMac(
-                        id: "mac-a",
-                        displayName: "Desk Mac",
-                        host: "100.82.214.112",
-                        lastSeenAt: Date(timeIntervalSince1970: 10),
-                        isActive: true
-                    ),
-                    try Self.pairedMac(
-                        id: "mac-b",
-                        displayName: "Laptop Mac",
-                        host: "100.82.214.113",
-                        lastSeenAt: Date(timeIntervalSince1970: 20),
-                        isActive: false
-                    ),
-                ],
-            ],
-            blockedTeams: []
-        )
-        let firstStore = MobileShellComposite(
-            isSignedIn: true,
-            pairedMacStore: pairedStore,
-            identityProvider: StaticIdentityProvider(userID: "user-1"),
-            teamIDProvider: { "team-a" },
-            forgottenMacStore: forgottenStore
-        )
-        await firstStore.loadPairedMacs()
-
-        await firstStore.forgetMac(macDeviceID: "mac-a")
-        try await pairedStore.upsert(
-            macDeviceID: "mac-a",
-            displayName: "Desk Mac",
-            routes: [try CmxAttachRoute(
-                id: "stale",
-                kind: .tailscale,
-                endpoint: .hostPort(host: "100.82.214.112", port: 50922)
-            )],
-            markActive: true,
-            stackUserID: "user-1",
-            teamID: "team-a",
-            now: Date(timeIntervalSince1970: 30)
-        )
-
-        let relaunchedStore = MobileShellComposite(
-            isSignedIn: true,
-            pairedMacStore: pairedStore,
-            identityProvider: StaticIdentityProvider(userID: "user-1"),
-            teamIDProvider: { "team-a" },
-            forgottenMacStore: forgottenStore
-        )
-
-        await relaunchedStore.loadPairedMacs()
-
-        #expect(relaunchedStore.pairedMacs.map(\.macDeviceID) == ["mac-b"])
-        #expect(relaunchedStore.displayPairedMacs.map(\.macDeviceID) == ["mac-b"])
-        #expect(await relaunchedStore.secondaryAggregationCandidateMacIDsForTesting() == ["mac-b"])
-    }
-
-    @Test func forgettingTeamlessMacSuppressesItAfterTeamSwitch() async throws {
-        let team = MutableTeamID("team-a")
-        let pairedStore = DelayedTeamPairedMacStore(
-            recordsByTeam: [
-                "": [
-                    try Self.pairedMac(
-                        id: "mac-legacy",
-                        displayName: "Legacy Mac",
-                        host: "100.82.214.112",
-                        lastSeenAt: Date(timeIntervalSince1970: 10),
-                        isActive: true,
-                        teamID: nil
-                    ),
-                ],
-                "team-b": [
-                    try Self.pairedMac(
-                        id: "mac-b",
-                        displayName: "Team B Mac",
-                        host: "100.82.214.113",
-                        lastSeenAt: Date(timeIntervalSince1970: 20),
-                        isActive: false,
-                        teamID: "team-b"
-                    ),
-                ],
-            ],
-            blockedTeams: []
-        )
-        let store = MobileShellComposite(
-            isSignedIn: true,
-            pairedMacStore: pairedStore,
-            identityProvider: StaticIdentityProvider(userID: "user-1"),
-            teamIDProvider: { await team.value },
-            forgottenMacStore: InMemoryPairedMacForgottenStore()
-        )
-        await store.loadPairedMacs()
-        #expect(store.displayPairedMacs.map(\.macDeviceID) == ["mac-legacy"])
-
-        await store.forgetMac(macDeviceID: "mac-legacy")
-        #expect(store.displayPairedMacs.isEmpty)
-
-        await team.set("team-b")
-        store.currentTeamDidChange()
-        await store.loadPairedMacs()
-
-        #expect(store.pairedMacs.map(\.macDeviceID) == ["mac-b"])
-        #expect(store.displayPairedMacs.map(\.macDeviceID) == ["mac-b"])
-    }
-
-    @Test func workspaceListReconnectPrefersSelectedUnavailableWorkspaceMac() async throws {
-        let pairedStore = DelayedTeamPairedMacStore(
-            recordsByTeam: [
-                "team-a": [
-                    try Self.pairedMac(
-                        id: "mac-a",
-                        displayName: "Desk Mac",
-                        host: "100.82.214.112",
-                        lastSeenAt: Date(timeIntervalSince1970: 10),
-                        isActive: true
-                    ),
-                    try Self.pairedMac(
-                        id: "mac-b",
-                        displayName: "Laptop Mac",
-                        host: "100.82.214.113",
-                        lastSeenAt: Date(timeIntervalSince1970: 20),
-                        isActive: false
-                    ),
-                ],
-            ],
-            blockedTeams: []
-        )
-        let store = MobileShellComposite(
-            isSignedIn: true,
-            pairedMacStore: pairedStore,
-            identityProvider: StaticIdentityProvider(userID: "user-1"),
-            teamIDProvider: { "team-a" }
-        )
-        await store.loadPairedMacs()
-        store.setWorkspaceStatesForTesting([
-            "mac-a": MacWorkspaceState(
-                macDeviceID: "mac-a",
-                workspaces: [
-                    MobileWorkspacePreview(
-                        id: "workspace-a",
-                        macDeviceID: "mac-a",
-                        name: "Desk",
-                        terminals: []
-                    ),
-                ],
-                status: .unavailable
-            ),
-            "mac-b": MacWorkspaceState(
-                macDeviceID: "mac-b",
-                workspaces: [
-                    MobileWorkspacePreview(
-                        id: "workspace-b",
-                        macDeviceID: "mac-b",
-                        name: "Laptop",
-                        terminals: []
-                    ),
-                ],
-                status: .unavailable
-            ),
-        ], foregroundMacDeviceID: "mac-a")
-
-        let workspaceB = try #require(store.workspaces.first {
-            $0.rpcWorkspaceID.rawValue == "workspace-b"
-        })
-        store.selectedWorkspaceID = workspaceB.id
-
-        #expect(store.workspaceListReconnectTargetMacDeviceID() == "mac-b")
-    }
-
-    @Test func workspaceListReconnectUsesSingleUnavailableWorkspaceOwner() async throws {
-        let pairedStore = DelayedTeamPairedMacStore(
-            recordsByTeam: [
-                "team-a": [
-                    try Self.pairedMac(
-                        id: "mac-b",
-                        displayName: "Laptop Mac",
-                        host: "100.82.214.113",
-                        lastSeenAt: Date(timeIntervalSince1970: 20),
-                        isActive: false
-                    ),
-                ],
-            ],
-            blockedTeams: []
-        )
-        let store = MobileShellComposite(
-            isSignedIn: true,
-            pairedMacStore: pairedStore,
-            identityProvider: StaticIdentityProvider(userID: "user-1"),
-            teamIDProvider: { "team-a" }
-        )
-        await store.loadPairedMacs()
-        store.setWorkspaceStatesForTesting([
-            "mac-b": MacWorkspaceState(
-                macDeviceID: "mac-b",
-                workspaces: [
-                    MobileWorkspacePreview(
-                        id: "workspace-b",
-                        macDeviceID: "mac-b",
-                        name: "Laptop",
-                        terminals: []
-                    ),
-                ],
-                status: .unavailable
-            ),
-        ], foregroundMacDeviceID: nil)
-
-        store.selectedWorkspaceID = nil
-
-        #expect(store.workspaceListReconnectTargetMacDeviceID() == "mac-b")
-    }
-
-    @Test func workspaceListReconnectDoesNotPickFirstUnavailableOwnerWithoutSelection() async throws {
-        let pairedStore = DelayedTeamPairedMacStore(
-            recordsByTeam: [
-                "team-a": [
-                    try Self.pairedMac(
-                        id: "mac-a",
-                        displayName: "Desk Mac",
-                        host: "100.82.214.112",
-                        lastSeenAt: Date(timeIntervalSince1970: 10),
-                        isActive: false
-                    ),
-                    try Self.pairedMac(
-                        id: "mac-b",
-                        displayName: "Laptop Mac",
-                        host: "100.82.214.113",
-                        lastSeenAt: Date(timeIntervalSince1970: 20),
-                        isActive: false
-                    ),
-                ],
-            ],
-            blockedTeams: []
-        )
-        let store = MobileShellComposite(
-            isSignedIn: true,
-            pairedMacStore: pairedStore,
-            identityProvider: StaticIdentityProvider(userID: "user-1"),
-            teamIDProvider: { "team-a" }
-        )
-        await store.loadPairedMacs()
-        store.setWorkspaceStatesForTesting([
-            "mac-a": MacWorkspaceState(
-                macDeviceID: "mac-a",
-                workspaces: [
-                    MobileWorkspacePreview(
-                        id: "workspace-a",
-                        macDeviceID: "mac-a",
-                        name: "Desk",
-                        terminals: []
-                    ),
-                ],
-                status: .unavailable
-            ),
-            "mac-b": MacWorkspaceState(
-                macDeviceID: "mac-b",
-                workspaces: [
-                    MobileWorkspacePreview(
-                        id: "workspace-b",
-                        macDeviceID: "mac-b",
-                        name: "Laptop",
-                        terminals: []
-                    ),
-                ],
-                status: .unavailable
-            ),
-        ], foregroundMacDeviceID: nil)
-
-        store.selectedWorkspaceID = nil
-
-        #expect(store.workspaceListReconnectTargetMacDeviceID() == nil)
-    }
-
-    @Test func workspaceListConnectedRefreshDoesNotPickFirstConnectedOwnerWithoutSelection() async throws {
-        let pairedStore = DelayedTeamPairedMacStore(
-            recordsByTeam: [
-                "team-a": [
-                    try Self.pairedMac(
-                        id: "mac-a",
-                        displayName: "Desk Mac",
-                        host: "100.82.214.112",
-                        lastSeenAt: Date(timeIntervalSince1970: 10),
-                        isActive: false
-                    ),
-                    try Self.pairedMac(
-                        id: "mac-b",
-                        displayName: "Laptop Mac",
-                        host: "100.82.214.113",
-                        lastSeenAt: Date(timeIntervalSince1970: 20),
-                        isActive: false
-                    ),
-                ],
-            ],
-            blockedTeams: []
-        )
-        let store = MobileShellComposite(
-            isSignedIn: true,
-            pairedMacStore: pairedStore,
-            identityProvider: StaticIdentityProvider(userID: "user-1"),
-            teamIDProvider: { "team-a" }
-        )
-        await store.loadPairedMacs()
-        store.setWorkspaceStatesForTesting([
-            "mac-a": MacWorkspaceState(
-                macDeviceID: "mac-a",
-                workspaces: [
-                    MobileWorkspacePreview(
-                        id: "workspace-a",
-                        macDeviceID: "mac-a",
-                        name: "Desk",
-                        terminals: []
-                    ),
-                ],
-                status: .connected
-            ),
-            "mac-b": MacWorkspaceState(
-                macDeviceID: "mac-b",
-                workspaces: [
-                    MobileWorkspacePreview(
-                        id: "workspace-b",
-                        macDeviceID: "mac-b",
-                        name: "Laptop",
-                        terminals: []
-                    ),
-                ],
-                status: .connected
-            ),
-        ], foregroundMacDeviceID: nil)
-
-        store.selectedWorkspaceID = nil
-
-        #expect(store.workspaceListConnectedRefreshTargetMacDeviceID() == nil)
-    }
-
     @Test func destructiveActionsDoNothingWithoutSignedInScope() async throws {
         let pairedStore = DelayedTeamPairedMacStore(
             recordsByTeam: [
@@ -1056,8 +352,7 @@ import Testing
         customName: String? = nil,
         customColor: String? = nil,
         customIcon: String? = nil,
-        routes: [CmxAttachRoute]? = nil,
-        teamID: String? = "team-a"
+        routes: [CmxAttachRoute]? = nil
     ) throws -> MobilePairedMac {
         MobilePairedMac(
             macDeviceID: id,
@@ -1067,7 +362,7 @@ import Testing
             lastSeenAt: lastSeenAt,
             isActive: isActive,
             stackUserID: "user-1",
-            teamID: teamID,
+            teamID: "team-a",
             customName: customName,
             customColor: customColor,
             customIcon: customIcon
